Rename generateRandomKey -> generateKey as it's no longer random
