+++ src/model/transaction/splitBlockInContentState.js
@@ -13,7 +13,7 @@
 
 'use strict';
 
-var generateRandomKey = require('generateRandomKey');
+var generateKey = require('generateKey');
 var invariant = require('invariant');
 
 import type ContentState from 'ContentState';
@@ -41,7 +41,7 @@ function splitBlockInContentState(
     characterList: chars.slice(0, offset),
   });
 
-  var keyBelow = generateRandomKey();
+  var keyBelow = generateKey();
   var blockBelow = blockAbove.merge({
     key: keyBelow,
     text: text.slice(offset),
